Refactor BinTree into LZ package.
