+++ lib/LZ.js
@@ -1,7 +1,8 @@
 if (typeof define !== 'function') { var define = require('amdefine')(module); }
-define(['./LZ/InWindow','./LZ/OutWindow'],function(InWindow,OutWindow){
+define(['./LZ/BinTree','./LZ/InWindow','./LZ/OutWindow'],function(BinTree,InWindow,OutWindow){
   'use strict';
   return {
+    BinTree: BinTree,
     InWindow: InWindow,
     OutWindow: OutWindow
   };
