+++ lib/LZ/InWindow.js
@@ -3,10 +3,12 @@ define(['../makeBuffer'],function(makeBuffer){
 'use strict';
 
 var InWindow = function(keepSizeBefore, keepSizeAfter, keepSizeReserve, stream){
-  if (arguments.length === 4) {
+  if (arguments.length >= 4) {
     // typical initialization sequence
-    this.create(keepSizeBefore, keepSizeAfter, keepSizeReserve);
-    this.setStream(stream);
+    var args = Array.prototype.slice.call(arguments, 0);
+    var _stream = args.pop();
+    this.create.apply(this, args);
+    this.setStream(_stream);
     this.init();
   }
 };
