+++ lib/encoder.js
@@ -1,5 +1,5 @@
 if (typeof define !== 'function') { var define = require('amdefine')(module); }
-define(['./rangeCoder','./bitEncoder','./binTree'], function(RangeCoder, BitEncoder, BinTree) {
+define(['./rangeCoder','./bitEncoder','./LZ'], function(RangeCoder, BitEncoder, LZ) {
 'use strict';
 
 function Encoder() {
@@ -361,7 +361,7 @@ function Encoder() {
                 var numHashBytes;
                 if (_matchFinder === null) {
                         numHashBytes = this._matchFinderType === 'BT2' ? 2 : 4;
-                        _matchFinder = new BinTree.BinTree();
+                        _matchFinder = new LZ.BinTree();
                         _matchFinder.setType(numHashBytes);
                 }
